Update BMD_FldKeyFreeAnnounce_Day_Evening.uasset
Updated mentions of art club in this file. Main reason for this being a smaller commit is just that I was using the file to test running proton on the command line instead of wine, and then I finally got it working and now I'm making sure this gets committed separately in case it goes wrong.